# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LifeOS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List subjects
      parameters:
        - in: query
          name: semester
          schema: { type: ["integer", "null"] }
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{id}:
    patch:
      operationId: updateSubject
      tags: [subjects]
      summary: Update a subject
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete a subject
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ASSIGNMENTS
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # BUCKET LIST
...
components:
  schemas:
...
    Subject:
      type: object
      required: [id, name, semester, credits]
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: ["string", "null"] }
        semester: { type: integer }
        credits: { type: number }
        gradePoint: { type: ["number", "null"] }
        grade: { type: ["string", "null"] }
    SubjectInput:
      type: object
      required: [name, semester, credits]
      properties:
        name: { type: string }
        code: { type: string }
        semester: { type: integer }
        credits: { type: number }
        gradePoint: { type: number }
        grade: { type: string }
    SubjectUpdate:
      type: object
      properties:
        name: { type: string }
        code: { type: string }
        semester: { type: integer }
        credits: { type: number }
        gradePoint: { type: number }
        grade: { type: string }
...
    Project:
      type: object
      required: [id, title, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        githubUrl: { type: ["string", "null"] }
        liveUrl: { type: ["string", "null"] }
        tags: { type: array, items: { type: string } }
        semester: { type: ["integer", "null"] }
        status: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        createdAt: { type: string }
    ProjectInput:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: string }
        githubUrl: { type: string }
        liveUrl: { type: string }
        tags: { type: array, items: { type: string } }
        semester: { type: integer }
        status: { type: string }
        imageUrl: { type: string }
    ProjectUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        githubUrl: { type: string }
        liveUrl: { type: string }
        tags: { type: array, items: { type: string } }
        semester: { type: integer }
        status: { type: string }
        imageUrl: { type: string }
...