# lib/api-zod/src/generated/types/subjectUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifeOS API
 * OpenAPI spec version: 0.1.0
 */

export interface SubjectUpdate {
  name?: string;
  code?: string;
  semester?: number;
  credits?: number;
  gradePoint?: number;
  grade?: string;
}